+++ components/GettingStarted/GettingStartedComponent/index.tsx
