--- pages/components/GettingStarted/GettingStartedSection/index.tsx
@@ -1,7 +1,0 @@
-import { Box } from "@chakra-ui/react";
-
-const GettingStartedSection = () => {
-  return <Box></Box>;
-};
-
-export default GettingStartedSection;
